substrate-adapter: move hashing test to seperate file
