+++ test/adapters/substrate/src/host_api/mod.rs
@@ -1,6 +1,7 @@
 mod allocator;
 mod child_storage;
 mod crypto;
+mod hashing;
 mod network;
 mod storage;
 mod trie;
@@ -54,13 +55,13 @@ pub fn process_host_api_tests(subcmd_matches: &ArgMatches) {
             "ext_crypto_sr25519_verify_version_1" => crypto::ext_crypto_sr25519_verify_version_1(rtm, input),
 
             // hashing api
-            "ext_hashing_keccak_256_version_1" => crypto::ext_hashing_keccak_256_version_1(rtm, input),
-            "ext_hashing_sha2_256_version_1" => crypto::ext_hashing_sha2_256_version_1(rtm, input),
-            "ext_hashing_blake2_128_version_1" => crypto::ext_hashing_blake2_128_version_1(rtm, input),
-            "ext_hashing_blake2_256_version_1" => crypto::ext_hashing_blake2_256_version_1(rtm, input),
-            "ext_hashing_twox_256_version_1" => crypto::ext_hashing_twox_256_version_1(rtm, input),
-            "ext_hashing_twox_128_version_1" => crypto::ext_hashing_twox_128_version_1(rtm, input),
-            "ext_hashing_twox_64_version_1" => crypto::ext_hashing_twox_64_version_1(rtm, input),
+            "ext_hashing_keccak_256_version_1" => hashing::ext_hashing_keccak_256_version_1(rtm, input),
+            "ext_hashing_sha2_256_version_1" => hashing::ext_hashing_sha2_256_version_1(rtm, input),
+            "ext_hashing_blake2_128_version_1" => hashing::ext_hashing_blake2_128_version_1(rtm, input),
+            "ext_hashing_blake2_256_version_1" => hashing::ext_hashing_blake2_256_version_1(rtm, input),
+            "ext_hashing_twox_256_version_1" => hashing::ext_hashing_twox_256_version_1(rtm, input),
+            "ext_hashing_twox_128_version_1" => hashing::ext_hashing_twox_128_version_1(rtm, input),
+            "ext_hashing_twox_64_version_1" => hashing::ext_hashing_twox_64_version_1(rtm, input),
 
             // allocator api
             "ext_allocator_malloc_version_1" => allocator::ext_allocator_malloc_version_1(rtm, input),
